pyproject.toml
[tool.poetry]
name = "mock-payment-api"
version = "0.1.0"
description = "mock payment api for learning DevOps"
authors = ["Jean Eric TUYISHIMIRE <ericjohn415@gmail.com>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "mock_payment_api"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.1"
uvicorn = {extras = ["standard"], version = "^0.23.1"}
sqlmodel = "^0.0.8"
psycopg2-binary = "^2.9.6"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
